+++ atlasdb-client/src/main/java/com/palantir/atlasdb/persister/JsonNodePersister.java
@@ -21,7 +21,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.palantir.atlasdb.persist.api.Persister;
-import com.palantir.atlasdb.processors.Reusable;
+import com.palantir.atlasdb.annotation.Reusable;
 import com.palantir.common.base.Throwables;
